+++ components/account/AccountList.tsx
@@ -118,7 +118,7 @@ const AccountList = () => {
 
   const handleChangeRowsPerPage = (event: React.ChangeEvent<HTMLInputElement>) => {
     setPageSize(+event.target.value);
-    setPage(0);
+    setPage(1);
   };
 
   const { loading, error, data } = useQuery(GET_ACCOUNTS, {
@@ -192,6 +192,13 @@ const AccountList = () => {
             page={page}
             onChangePage={handleChangePage}
             onChangeRowsPerPage={handleChangeRowsPerPage}
+            backIconButtonProps={{
+              'aria-label': 'Previous',
+              'disabled': page === 1,
+            }}
+            nextIconButtonProps={{
+              'aria-label': 'Next',
+            }}
           />
         </Paper>
       </Grid>