+++ pages/block/[block].tsx
@@ -23,7 +23,7 @@ export default function Block() {
   return (
     <Layout>
       <Grid container spacing={2} className={classes.root}>
-        <Grid item xs={12} sm={8} >
+        <Grid item xs={12} md={7} >
           <BlockDetails />
         </Grid>
       </Grid>